Initial configuration for the admin panel
