+++ gateway/admin_panel/admin.py
@@ -1,0 +1,27 @@
+from django.contrib import admin
+
+from api.models import JobConfig, Provider, Program, ComputeResource, Job, RuntimeJob
+
+@admin.register(JobConfig)
+class JobConfigAdmin(admin.ModelAdmin):
+    """JobConfigAdmin."""
+
+@admin.register(Provider)
+class ProviderAdmin(admin.ModelAdmin):
+    """ProviderAdmin."""
+
+@admin.register(Program)
+class ProgramAdmin(admin.ModelAdmin):
+    """ProgramAdmin."""
+
+@admin.register(ComputeResource)
+class ComputeResourceAdmin(admin.ModelAdmin):
+    """ComputeResourceAdmin."""
+
+@admin.register(Job)
+class JobAdmin(admin.ModelAdmin):
+    """JobAdmin."""
+
+@admin.register(RuntimeJob)
+class RuntimeJobAdmin(admin.ModelAdmin):
+    """RuntimeJobAdmin."""
