--- gateway/api/admin.py
@@ -1,19 +1,0 @@
-"""Admin module."""
-
-from django.contrib import admin
-from .models import Job, Program, ComputeResource
-
-
-@admin.register(Program)
-class ProgramAdmin(admin.ModelAdmin):
-    """ProgramAdmin."""
-
-
-@admin.register(Job)
-class JobAdmin(admin.ModelAdmin):
-    """JobAdmin."""
-
-
-@admin.register(ComputeResource)
-class ComputeResourceAdmin(admin.ModelAdmin):
-    """ComputeResourceAdmin."""
